+++ traffic_ops/testing/api/v2/withobjs_test.go
@@ -66,7 +66,6 @@ const (
 	SteeringTargets
 	Tenants
 	ServerCheckExtensions
-	Topologies
 	Types
 	Users
 )
@@ -107,7 +106,6 @@ var withFuncs = map[TCObj]TCObjFuncs{
 	SteeringTargets:                      {SetupSteeringTargets, DeleteTestSteeringTargets},
 	Tenants:                              {CreateTestTenants, DeleteTestTenants},
 	ServerCheckExtensions:                {CreateTestServerCheckExtensions, DeleteTestServerCheckExtensions},
-	Topologies:                           {CreateTestTopologies, DeleteTestTopologies},
 	Types:                                {CreateTestTypes, DeleteTestTypes},
 	Users:                                {CreateTestUsers, ForceDeleteTestUsers},
 }