+++ traffic_ops/testing/api/v3/topologies_test.go
@@ -1,4 +1,4 @@
-package v2
+package v3
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one