+++ traffic_ops/testing/api/v3/withobjs_test.go
@@ -66,6 +66,7 @@ const (
 	SteeringTargets
 	Tenants
 	ServerCheckExtensions
+	Topologies
 	Types
 	Users
 )
@@ -106,6 +107,7 @@ var withFuncs = map[TCObj]TCObjFuncs{
 	SteeringTargets:                      {SetupSteeringTargets, DeleteTestSteeringTargets},
 	Tenants:                              {CreateTestTenants, DeleteTestTenants},
 	ServerCheckExtensions:                {CreateTestServerCheckExtensions, DeleteTestServerCheckExtensions},
+	Topologies:                           {CreateTestTopologies, DeleteTestTopologies},
 	Types:                                {CreateTestTypes, DeleteTestTypes},
 	Users:                                {CreateTestUsers, ForceDeleteTestUsers},
 }